+++ core/services/workflows/models.go
@@ -1,48 +1,13 @@
 package workflows
 
 import (
-	"errors"
 	"fmt"
 
 	"github.com/dominikbraun/graph"
 
 	"github.com/smartcontractkit/chainlink-common/pkg/capabilities"
 	"github.com/smartcontractkit/chainlink-common/pkg/values"
-	"github.com/smartcontractkit/chainlink/v2/core/services/workflows/store"
 )
-
-type stepRequest struct {
-	stepRef string
-	state   store.WorkflowExecution
-}
-
-// StepDefinition is the parsed representation of a step in a workflow.
-//
-// Within the workflow spec, they are called "Capability Properties".
-type StepDefinition struct {
-	ID     string         `json:"id" jsonschema:"required"`
-	Ref    string         `json:"ref,omitempty" jsonschema:"pattern=^[a-z0-9_]+$"`
-	Inputs map[string]any `json:"inputs,omitempty"`
-	Config map[string]any `json:"config" jsonschema:"required"`
-
-	CapabilityType capabilities.CapabilityType `json:"-"`
-}
-
-// WorkflowSpec is the parsed representation of a workflow.
-type WorkflowSpec struct {
-	Triggers  []StepDefinition `json:"triggers" jsonschema:"required"`
-	Actions   []StepDefinition `json:"actions,omitempty"`
-	Consensus []StepDefinition `json:"consensus" jsonschema:"required"`
-	Targets   []StepDefinition `json:"targets" jsonschema:"required"`
-}
-
-func (w *WorkflowSpec) Steps() []StepDefinition {
-	s := []StepDefinition{}
-	s = append(s, w.Actions...)
-	s = append(s, w.Consensus...)
-	s = append(s, w.Targets...)
-	return s
-}
 
 // workflow is a directed graph of nodes, where each node is a step.
 //
@@ -114,133 +79,18 @@ type step struct {
 	executionStrategy executionStrategy
 }
 
-type Vertex struct {
-	StepDefinition
-	dependencies []string
-}
-
-// DependencyGraph is an intermediate representation of a workflow wherein all the graph
-// vertices are represented and validated. It is a static representation of the workflow dependencies.
-type DependencyGraph struct {
-	ID string
-	graph.Graph[string, *Vertex]
-
-	Triggers []*StepDefinition
-
-	Spec *WorkflowSpec
-}
-
-// VID is an identifier for a Vertex that can be used to uniquely identify it in a graph.
-// it represents the notion `hash` in the graph package AddVertex method.
-// we refrain from naming it `hash` to avoid confusion with the hash function.
-func (v *Vertex) VID() string {
-	return v.Ref
-}
-
 type triggerCapability struct {
 	StepDefinition
 	trigger capabilities.TriggerCapability
 	config  *values.Map
 }
 
-const (
-	keywordTrigger = "trigger"
-)
-
 func Parse(yamlWorkflow string) (*workflow, error) {
-	wf2, err := ParseDepedencyGraph(yamlWorkflow)
+	wf2, err := ParseDependencyGraph(yamlWorkflow)
 	if err != nil {
 		return nil, err
 	}
 	return createWorkflow(wf2)
-}
-
-func ParseDepedencyGraph(yamlWorkflow string) (*DependencyGraph, error) {
-	spec, err := ParseWorkflowSpecYaml(yamlWorkflow)
-	if err != nil {
-		return nil, err
-	}
-
-	// Construct and validate the graph. We instantiate an
-	// empty graph with just one starting entry: `trigger`.
-	// This provides the starting point for our graph and
-	// points to all dependent steps.
-	// Note: all triggers are represented by a single step called
-	// `trigger`. This is because for workflows with multiple triggers
-	// only one trigger will have started the workflow.
-	stepHash := func(s *Vertex) string {
-		return s.VID()
-	}
-	g := graph.New(
-		stepHash,
-		graph.PreventCycles(),
-		graph.Directed(),
-	)
-	err = g.AddVertex(&Vertex{
-		StepDefinition: StepDefinition{Ref: keywordTrigger},
-	})
-	if err != nil {
-		return nil, err
-	}
-
-	// Next, let's populate the other entries in the graph.
-	for _, s := range spec.Steps() {
-		// TODO: The workflow format spec doesn't always require a `Ref`
-		// to be provided (triggers and targets don't have a `Ref` for example).
-		// To handle this, we default the `Ref` to the type, but ideally we
-		// should find a better long-term way to handle this.
-		if s.Ref == "" {
-			s.Ref = s.ID
-		}
-
-		innerErr := g.AddVertex(&Vertex{StepDefinition: s})
-		if innerErr != nil {
-			return nil, fmt.Errorf("cannot add vertex %s: %w", s.Ref, innerErr)
-		}
-	}
-
-	stepRefs, err := g.AdjacencyMap()
-	if err != nil {
-		return nil, err
-	}
-
-	// Next, let's iterate over the steps and populate
-	// any edges.
-	for stepRef := range stepRefs {
-		step, innerErr := g.Vertex(stepRef)
-		if innerErr != nil {
-			return nil, innerErr
-		}
-
-		refs, innerErr := findRefs(step.Inputs)
-		if innerErr != nil {
-			return nil, innerErr
-		}
-		step.dependencies = refs
-
-		if stepRef != keywordTrigger && len(refs) == 0 {
-			return nil, errors.New("all non-trigger steps must have a dependent ref")
-		}
-
-		for _, r := range refs {
-			innerErr = g.AddEdge(r, step.Ref)
-			if innerErr != nil {
-				return nil, innerErr
-			}
-		}
-	}
-
-	triggerSteps := []*StepDefinition{}
-	for _, t := range spec.Triggers {
-		tt := t
-		triggerSteps = append(triggerSteps, &tt)
-	}
-	wf := &DependencyGraph{
-		Spec:     &spec,
-		Graph:    g,
-		Triggers: triggerSteps,
-	}
-	return wf, err
 }
 
 // createWorkflow converts a StaticWorkflow to an executable workflow
